Removed @xobotyi/scrollbar-width and added dom-helpers/scrollbarSize Refactored import paths of the new keyboard hook
