+++ src/editorManager/__tests__/useEditorManager.spec.tsx
@@ -1,13 +1,8 @@
 import { renderHook, act } from '@testing-library/react-hooks'
 import { useEditorManager } from '../useEditorManager'
 import { createColumnMock } from '../../columnGrid/__mocks__/column-mock'
-import { ROW_SELECTION_HEADER_ID, useRowSelection } from '../../rowSelection/useRowSelection'
-import TextEditor from '../components/TextEditor'
-import { EditorProps } from '../editorProps'
+import {  useRowSelection } from "../../rowSelection"
 import React, { useRef } from 'react'
-import { isElementOfType } from 'react-dom/test-utils'
-import { NavigationCoords } from '../../navigation/types/navigation-coords.type'
-import { ICellEditor } from '../../columnGrid/types/header.type'
 import { useApiFactory, useApiRef } from '../../api'
 import { useData } from '../../data'
 import { useHeaders } from '../../columnGrid'
