+++ src/index.tsx
@@ -6,7 +6,7 @@ export * from './data'
 export * from './editorManager'
 export * from './gridWrapper'
 export * from './mergeCells'
-export * from './navigation'
+export * from './keyboard'
 export * from './rowSelection'
 export * from './nestedRows'
 export * from './types'
