+++ src/mergeCells/validationErrorMessages.ts
@@ -1,4 +1,4 @@
-import { NavigationCoords } from '../navigation/types/navigation-coords.type'
+import { NavigationCoords } from '../keyboard/types/navigation-coords.type'
 
 export const MERGED_NEGATIVE_VALUES = ({ rowIndex, colIndex }: NavigationCoords) =>
   `Merged cell contains negative values, therefore it cannot be added to the collection [${rowIndex},${colIndex}]`
